[core] Merge Sprite into SpriteStore
